requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
pydantic==1.10.2
PyMySQL==1.0.2
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
